tools/music/requirements.txt
# Core async and HTTP libraries
asyncio
aiohttp>=3.8.0
aiofiles>=0.8.0

# Spotify API integration
spotipy>=2.22.0

# YouTube API integration
google-api-python-client>=2.70.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Audio processing and speech recognition
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10
pyaudio>=0.2.11
scipy>=1.7.0
numpy>=1.21.0

# Natural language processing
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0
nltk>=3.8.0

# Data processing and utilities
pandas>=1.5.0
pydantic>=1.10.0
python-dateutil>=2.8.0
pytz>=2022.1

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=22.1.0
colorlog>=6.7.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Development tools (optional)
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Additional audio format support
mutagen>=1.45.0  # For metadata extraction
pydub>=0.25.0    # For audio format conversion

# Machine learning models (optional)
scikit-learn>=1.1.0
matplotlib>=3.5.0  # For visualizations

# Web scraping (fallback for music data)
beautifulsoup4>=4.11.0
requests>=2.28.0

# Database support (for caching)
aiosqlite>=0.17.0
redis>=4.3.0  # Optional Redis caching

# Security and encryption
cryptography>=37.0.0
keyring>=23.0.0  # For secure credential storage

# Performance optimization
uvloop>=0.17.0  # Fast async event loop (Unix only)
orjson>=3.8.0   # Fast JSON parsing

# CLI enhancements
click>=8.1.0
rich>=12.5.0    # For beautiful terminal output
tqdm>=4.64.0    # Progress bars

# File format support
python-magic>=0.4.27  # File type detection
pillow>=9.2.0         # Image processing for thumbnails

# Monitoring and metrics
prometheus-client>=0.14.0
psutil>=5.9.0

# Environment compatibility
python-multipart>=0.0.5
httpx>=0.23.0 